Add optional uname kernel check for non-ape loader
ELF executables of the same architecture lack mechanisms to specify the
host operating system they are built for. As an example, SerenityOS and
Linux both specify SYSV in the OSABI field, and while Linux specifies a
.note.ABI-tag, SerenityOS lacks one altogether. Therefore, we do not do
autodetection of the target kernel from provided loaders, and defer the
matching of loaders to uname -s to the user.